move module definitions and remove empty cases
