--- test/beaver_test.exs
@@ -1,8 +1,0 @@
-defmodule BeaverTest do
-  use ExUnit.Case
-  doctest Beaver
-
-  test "greets the world" do
-    assert true
-  end
-end
