--- test/dialect_test.exs
@@ -1,7 +1,0 @@
-defmodule Beaver.MLIR.DialectTest do
-  use ExUnit.Case
-  doctest Beaver.MLIR.Dialect
-
-  test "greets the world" do
-  end
-end
